Create a different folder for firebase database
